refactor: rename src/classes to src/api
