+++ src/api/common/attack-object.impl.ts
