+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from '@/classes/index.js';
+export * from '@/api/index.js';
 export * from '@/data-sources/index.js';
 export * from '@/schemas/index.js';
 export * from '@/refinements/index.js';
